Update project settings and code style
This commit adds and modifies a number of project setting files. It introduces a new vcs.xml file, setting up Git for the project, and a .name file for the printer utility project. Changes have also been made to the editor settings regarding the CppFormatting. Lastly, this update modifies the encodings.xml, modifying charset paths to include library files in new subdirectories.